# api/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Rinna API",
    "description": "API for the Rinna project management system\n\n## Security\n\nThe Rinna API implements several security features to protect your data:\n\n- API Key Authentication: All endpoints require a valid API key in the Authorization header\n- Rate Limiting: Configurable request limits to prevent abuse\n- OAuth Integration: Connect to third-party services securely\n- Webhook Security: End-to-end verification of webhook payloads\n- Request Tracing: All requests include a unique X-Request-ID for traceability\n\n## Rate Limiting\n\nRate limiting is applied to all API endpoints to ensure fair usage and system stability. When rate limited, you will receive a 429 Too Many Requests response with a Retry-After header.\n\nRate limit headers:\n- X-RateLimit-Limit: Maximum requests allowed per minute\n- X-RateLimit-Remaining: Remaining requests in the current window\n- X-RateLimit-Reset: Timestamp when the limit resets (Unix time)\n\n## Detailed Documentation\n\nFor more detailed documentation, please refer to:\n- [API Security Guide](./docs/api-security-guide.md)\n- [Rate Limiting Documentation](./docs/rate-limiting.md)\n- [Secure Integration Examples](./docs/secure-integration-examples.md)\n",
    "version": "1.6.6",
    "contact": {
      "name": "Eric C. Mumford",
      "email": "info@example.com"
    },
    "license": {
      "name": "MIT"
    }
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "schemes": [
...
  "consumes": [
...
  "produces": [
...
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "API key authentication using the Bearer scheme. Example: 'Authorization: Bearer ri-api-key'"
    },
    "OAuth2": {
      "type": "oauth2",
      "flow": "accessCode",
      "authorizationUrl": "https://localhost:8080/api/v1/oauth/authorize",
      "tokenUrl": "https://localhost:8080/api/v1/oauth/token",
      "scopes": {
        "read": "Read access to projects and work items",
        "write": "Write access to projects and work items",
        "admin": "Administrative access"
      }
    }
  },
  "paths": {
...
    "/workitems": {
      "get": {
        "tags": [
          "workitems"
        ],
        "summary": "List all work items",
        "description": "Returns a list of work items with optional pagination and filtering",
        "operationId": "listWorkItems",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter by work item status",
            "required": false,
            "type": "string",
            "enum": [
              "FOUND",
              "TRIAGED",
              "IN_DEV",
              "TESTING",
              "DONE",
              "CLOSED"
            ]
          },
          {
            "name": "priority",
            "in": "query",
            "description": "Filter by priority",
            "required": false,
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ]
          },
          {
            "name": "type",
            "in": "query",
            "description": "Filter by work item type",
            "required": false,
            "type": "string",
            "enum": [
              "BUG",
              "FEATURE",
              "CHORE"
            ]
          },
          {
            "name": "project",
            "in": "query",
            "description": "Filter by project key",
            "required": false,
            "type": "string"
          },
          {
            "name": "release",
            "in": "query",
            "description": "Filter by release ID",
            "required": false,
            "type": "string"
          },
          {
            "name": "assignee",
            "in": "query",
            "description": "Filter by assignee",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "size",
            "in": "query",
            "description": "Number of items per page",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/WorkItem"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "429": {
            "description": "Too Many Requests",
            "headers": {
              "Retry-After": {
                "type": "integer",
                "description": "Seconds to wait before retrying"
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": [
          "workitems"
        ],
        "summary": "Create a new work item",
        "description": "Creates a new work item with the provided information",
        "operationId": "createWorkItem",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Work item object to be created",
            "required": true,
            "schema": {
              "$ref": "#/definitions/WorkItemCreateRequest"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Work item created successfully",
            "schema": {
              "$ref": "#/definitions/WorkItem"
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "429": {
            "description": "Too Many Requests",
            "headers": {
              "Retry-After": {
                "type": "integer",
                "description": "Seconds to wait before retrying"
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/webhooks/github": {
      "post": {
        "tags": [
          "webhooks"
        ],
        "summary": "GitHub webhook handler",
        "description": "Handles incoming GitHub events. Includes security validation with HMAC-SHA256 signatures.\nRequires X-Hub-Signature-256 header for webhook payload verification.\n",
        "operationId": "handleGithubWebhookV1",
        "parameters": [
          {
            "name": "X-GitHub-Event",
            "in": "header",
            "description": "GitHub event type",
            "required": true,
            "type": "string"
          },
          {
            "name": "X-Hub-Signature-256",
            "in": "header",
            "description": "GitHub signature for payload verification (sha256=HMAC signature)",
            "required": true,
            "type": "string"
          },
          {
            "name": "X-GitHub-Delivery",
            "in": "header",
            "description": "Unique identifier for the webhook delivery (used for nonce verification)",
            "required": true,
            "type": "string"
          },
          {
            "name": "project",
            "in": "query",
            "description": "Project key for the webhook target",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "GitHub event payload",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GitHubEvent"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Event processed successfully"
          },
          "400": {
            "description": "Invalid event payload"
          },
          "401": {
            "description": "Invalid signature or authentication failure"
          },
          "429": {
            "description": "Rate limit exceeded",
            "headers": {
              "Retry-After": {
                "type": "integer",
                "description": "Seconds to wait before retrying"
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/webhooks/gitlab": {
      "post": {
        "tags": [
          "webhooks"
        ],
        "summary": "GitLab webhook handler",
        "description": "Handles incoming GitLab events. Includes security validation with token verification.\nRequires X-Gitlab-Token header for webhook payload verification.\n",
        "operationId": "handleGitlabWebhook",
        "parameters": [
          {
            "name": "X-Gitlab-Event",
            "in": "header",
            "description": "GitLab event type",
            "required": true,
            "type": "string"
          },
          {
            "name": "X-Gitlab-Token",
            "in": "header",
            "description": "GitLab token for payload verification",
            "required": true,
            "type": "string"
          },
          {
            "name": "project",
            "in": "query",
            "description": "Project key for the webhook target",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "GitLab event payload",
            "required": true,
            "schema": {
              "type": "object"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Event processed successfully"
          },
          "400": {
            "description": "Invalid event payload"
          },
          "401": {
            "description": "Invalid token or authentication failure"
          },
          "429": {
            "description": "Rate limit exceeded",
            "headers": {
              "Retry-After": {
                "type": "integer",
                "description": "Seconds to wait before retrying"
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/webhooks/bitbucket": {
      "post": {
        "tags": [
          "webhooks"
        ],
        "summary": "Bitbucket webhook handler",
        "description": "Handles incoming Bitbucket events. Includes security validation with HMAC-SHA1 signatures.\nRequires X-Hub-Signature header for webhook payload verification.\n",
        "operationId": "handleBitbucketWebhook",
        "parameters": [
          {
            "name": "X-Event-Key",
            "in": "header",
            "description": "Bitbucket event type",
            "required": true,
            "type": "string"
          },
          {
            "name": "X-Hub-Signature",
            "in": "header",
            "description": "Bitbucket signature for payload verification",
            "required": true,
            "type": "string"
          },
          {
            "name": "X-Request-UUID",
            "in": "header",
            "description": "Unique identifier for the webhook delivery (used for nonce verification)",
            "required": true,
            "type": "string"
          },
          {
            "name": "project",
            "in": "query",
            "description": "Project key for the webhook target",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Bitbucket event payload",
            "required": true,
            "schema": {
              "type": "object"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Event processed successfully"
          },
          "400": {
            "description": "Invalid event payload"
          },
          "401": {
            "description": "Invalid signature or authentication failure"
          },
          "429": {
            "description": "Rate limit exceeded",
            "headers": {
              "Retry-After": {
                "type": "integer",
                "description": "Seconds to wait before retrying"
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "WorkItem": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uuid"
        },
        "title": {
          "type": "string",
          "example": "Fix login button on home page"
        },
        "description": {
          "type": "string"
        },
        "type": {
          "type": "string",
          "enum": [
            "BUG",
            "FEATURE",
            "CHORE"
          ]
        },
        "priority": {
          "type": "string",
          "enum": [
            "LOW",
            "MEDIUM",
            "HIGH"
          ]
        },
        "status": {
          "type": "string",
          "enum": [
            "FOUND",
            "TRIAGED",
            "IN_DEV",
            "TESTING",
            "DONE",
            "CLOSED"
          ]
        },
        "assignee": {
          "type": "string"
        },
        "projectKey": {
          "type": "string",
          "example": "RINNA"
        },
        "releaseId": {
          "type": "string"
        },
        "metadata": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
  }
}